fix(engine): engine and context core fixes
- COW 就地排序数据竞争修复: 排序前做逐元素完整拷贝
- WithSharedExecPool 失效修复: 仅在未注入时创建自有池
- 中间件热更新不达临时 matcher 修复: tempManager.ForEach 统一失效
- OnTemp 后设超时永不过期修复: SetExpiration shard 锁内写入+补登堆
- DeleteMatcher 恢复批量删除路径: 新增 FlushPendingDeletes
- 链式元数据缓存陈旧修复: Set*/BindCommand 触发 UpdateCommandCache
- 批量注册命令元数据缺失修复: withBatchMatchers 更新 commandInfoCache
- 命令索引分类收紧: commandIndexed 标志判定命令 matcher
- Dispatcher 同 chat 双 worker 竞态修复: 持锁复查映射并重试
- ExecPool 队列滞留修复: exitMu 退出协议
- 归并迭代器哨兵值边界修复: 显式 found 判定 + uint64 全程比较
- TempManager 快照写放大修复: COW 增量插入/移除
- Context DeferRuleEffect/Commit/Discard 延迟副作用机制
- OnMentionedBot 平台不支持时永不触发修复
- 温和修复: StartTempMatcherCleaner writeMu, cleanupWg Shutdown 等待

+++ core/engine/matcher_test.go
@@ -73,7 +73,7 @@ func TestMatcher_CopySemantics(t *testing.T) {
 	t.Run("nil matcher nil-safety", func(t *testing.T) {
 		assert.Equal(t, "", (*Matcher)(nil).GetCommand())
 		assert.Equal(t, "", (*Matcher)(nil).GetSource())
-		assert.Equal(t, uint(0), (*Matcher)(nil).getPriority())
+		assert.Equal(t, uint64(0), (*Matcher)(nil).getPriority())
 		assert.False(t, (*Matcher)(nil).isBlocking(""))
 	})
 }
